feat: Update getOneUserService to handle userIdentifier as ObjectId or name
This commit updates the getOneUserService function in the getOneUser module to handle the userIdentifier parameter as either an ObjectId or a name. If the userIdentifier is a valid ObjectId, the function will search for a user by ID using the findById method. If the userIdentifier is not a valid ObjectId, the function will search for a user by name using the findOne method. This change allows for more flexible user lookup based on different types of identifiers.

Recent repository commits:
- feat: Update User model to include permissions as enum type
- feat: Update KanbanColumn model to include boardId and taskIds as ObjectId types
- feat: Update KanbanTaskSchema to include files array
- feat: Update getAllNotificationsService to use Notifications model
- feat: Update updateNotificationService to updateNotificationController
- feat: Update getAllNotificationsController to use getAllNotificationsService
- feat: Remove 'finished' field from Notifications model
- feat: Update Notifications model to remove 'finished' field