feat: Update package.json and config.ts files
This commit updates the package.json file to include the "@types/node" dependency with version "^22.1.0". This update ensures compatibility with the latest version of Node.js.

Additionally, the config.ts file is updated to change the HOST_API variable. In the production environment, the HOST_API is set to "http://192.168.2.15", and in the development environment, it is set to "http://192.168.2.15:8000". This change ensures that the correct API host is used based on the environment.

Recent repository commits:
- feat: Update KanbanColumn model to include boardId and taskIds as ObjectId types
- feat: Update getOneUserService to handle userIdentifier as ObjectId or name
- feat: Update User model to include permissions as enum type
- feat: Update KanbanTaskSchema to include files array
- feat: Update getAllNotificationsService to use Notifications model
- feat: Update updateNotificationService to updateNotificationController
- feat: Update getAllNotificationsController to use getAllNotificationsService
- feat: Remove 'finished' field from Notifications model
- feat: Update Notifications model to remove 'finished' field