feat: Update package.json and config.ts files
This commit updates the package.json file to include the "@types/node" dependency with version "^22.1.0". This update ensures compatibility with the latest version of Node.js.

Additionally, the config.ts file is updated to change the HOST_API variable. In the production environment, the HOST_API is set to "http://192.168.2.15", and in the development environment, it is set to "http://192.168.2.15:8000". This change ensures that the correct API host is used based on the environment.